requirements.txt
torch>=2.0.0
torchvision
transformers==4.35.0
timm
omegaconf
pytorch-lightning==1.9.5
kornia
webdataset
albumentations==1.3.1
scipy
numpy
Pillow
matplotlib
gradio
fastapi
uvicorn
lpips
torch-fidelity